Issue 106: Handling of large or empty document content.
This change forces document content that exceeds the maxDocumentSize
to be discard and replaced with the default placeholder content
(the google:title value or a space).  This mirrors the behavior
of Connectors that implement TraversalContextAware and choose
to not supply a content stream for documents that exceed the
TraversalContext.maxDocumentSize.

As a consequence of this change, I also cleared up a TODO:
Documents with non-null, but empty, content streams also
provide the default placeholder content.

The mechanics of the change are simple.  DocPusher has
two FilterInputStreams that sit in the document content
pipeline:
 - BigEmptyDocumentFilterInputStream.  Sits right next to
   the Connector's provided content input stream.  If reads
   to that input stream return 0 bytes the filter throws
   EmptyDocumentException (a subclass of IOException).
   If the number of bytes read from the content exceed the
   maximum document size, the filter throws EmptyDocumentException.

 - AlternateContentFilterInputStream wraps the full
   content read, including Base64Encoding, Compression,
   and the content stream.  It takes as a parameter
   an alternate content stream to provide if it catches
   EmptyDocumentException or BigDocumentException.
   If it catches one of those exceptions, it resets
   the accumulated xmlFeed write position to the start
   of the content section, and switches to providing
   the alternate content on reads.