This change fixes an error in handling big documents.
The code for supplying the alternate content (title or space)
for documents that exceed the maximum document size was broken.
The alternate content was the right size, but was read into the
wrong location of the I/O buffer.

Files Modified:
M  projects/connector-manager/source/java/com/google/enterprise/connector/pusher/DocPusher.java
   - AlternateContentFilterInputStream.read(byte[], int, int)
     now returns zero bytes read after switching to the
     alternate content stream.  The alternate content is then
     returned on subsequent calls to read.