Issue 143 - Adds SkippedDocumentException for support of 'ignored' mimetypes.
Revision r2281 adds a new MimeTypeMap set for documents
that should be 'ignored' (skipped).  Basically the use is:

If TraversalContext.mimeTypeSupportLevel(String mimetype)
returns -1, the document should be skipped.  Do not index
metadata. Do not index content.

At this time, support for this feature still rests with the
connector implementation.  John and I were discussing the
best way to implement skipping documents that fall into the
'ignored' set.  The connector could skip them internally -
either silently or log a message.  This would be easy for
the Livelink connector, but difficult for the Documentum
connector.

Months ago, Jeff and Marty mentioned that customers wanted
to know which documents were skipped and for what reason.
From that perspective, it makes more sense for the Connector
to throw a RepositoryDocumentException, which will skip the
document and log the action.  From the implementation point
of view, it is already supported by the CM, and could be
easily implemented in the Livelink and Documentum Connectors
with 2-4 lines of code.

One downside is that RepositoryDocumentExceptions are
logged at WARNING level.  This change creates a new subclass
of RepositoryDocumentException, SkipDocumentException,
that gets caught separately and logged at FINER.  Connectors
could use it not just for 'ignored' mimetypes, but any other
benevolent reason that a document gets skipped (perhaps some
other advanced config option).