Fix for Issue 204.
Note the stack trace pasted into the issue report is actually a problem with the
customer's Connector.  It appears the Connector is not properly escaping the
config values when putting together the formSnippet.  See Issue 186 for details.

However, Issue 204 exposed a problem when the CDATA end marker ']]>' is actually
embedded in the formSnippet.  Because the formSnippet is surrounded by a CDATA
section, this can cause the Admin Console to truncate the XML for the
formSnippet causing an incomplete element to be extracted and a
NullPointerException on the Search Appliance.


Tests Passed:

  - All Unit tests.
  - Tested with TestConnector and GSA using string from the issue report.
    String was returned when Connector was edited.


ChangeLog:

  M java/com/google/enterprise/connector/servlet/ServletUtil.java
    - Fixed the removeNestedMarkers() utility to only work with pairs of
      markers. TODO, this could still remove the wrong end marker since the
      matching of the end marker doesn't work from the back of the form.
    - Added another utility to escape the CDATA end marker so it won't be mistaken
      for the real end marker when the form is wrapped.

  M java/com/google/enterprise/connector/servlet/ConnectorManagerGetServlet.java
    - Added additional call to escape any CDATA end markers so they don't truncate
      the form on the Search Appliance side.

  M javatests/com/google/enterprise/connector/servlet/ServletUtilTest.java
    - Added related tests.