Issue 206 - Auto-Disabled connectors cannot be re-enabled
Once a run-once connector disabled itself, the HostLoadManager got an
entry for it in its connectorNameToFinishTime table that set a finish
time of infinity.  Re-enabling the connector did not clear this entry,
so the connector would never get scheduled to run.

This change sets the connectorNameToFinishTime entry to 1 minute rather
than infinity.  This gives the ConnectorCoordinator enough time to
save the modified schedule without worry the connector will get
rescheduled, but won't impact some future re-enabling of the schedule.
Once the disabled schedule is saved, the connector will not be scheduled
for a traversal because of its disabled state.

Change Log:
M  projects/connector-manager/source/java/com/google/enterprise/connector/scheduler/HostLoadManager.java
   - In connectorFinishedTraversal(), change the timeout for negative
     parameters from Long.MAX_VALUE to 60,000.