This issue is urgent because we may judge that it needs to go into 6.8.
The substantive change is the addition of SpiConstants.PROPNAME_LOCK and special
handling for that property such that if it is true, then the lock=true attribute
is added to the record in the feed document.

This should go into head in any case - we should be prepared to DI into 2.6.2
this afternoon if they decide they want it. 

Note: I have not yet done and end-to-end test with a GSA showing that this works
with the GSA, so please hold-off on the DI until I can report on testing.

http://codereview.appspot.com/2236041/