Add SkipDocumentFilter, a document filter that throws SkipDocumentExceptions based upon presence/absence of specific metadata.  The filter has three configuration properties:
propertyName - the property name to filter upon (required).

pattern - a regular expression to match on the property's values (optional).

skipOnMatch - a boolean (default: false). If true, skip the document if it
    matches the criteria. If false, skip the document if it fails to match
    the criteria.

If pattern is set, it is applied to each of the property's values.  If a
value matches the regular expression, the document is considered a match.
If pattern is not set, the mere presence of the property would be considered
a match.

Request for comment: Should the default for skipOnMatch be true or false?

Request for comment: Should the pattern be an exact match on the value,
or simply match a region of the value?  The current implementation does
the latter (as does the ModifyPropertyFilter).

Change Log:
----------
A  projects/connector-manager/source/java/com/google/enterprise/connector/util/filter/SkipDocumentFilter.java
M  projects/connector-manager/source/javatests/com/google/enterprise/connector/util/filter/DocumentFilterTest.java
A  projects/connector-manager/source/javatests/com/google/enterprise/connector/util/filter/SkipDocumentFilterTest.java

Code Review:
-----------
http://codereview.appspot.com/4521044