Allow certificate security for GetDocumentContent
The client accessing GetDocumentContent needs to provide a client
certificate. The Common Name of the Subject of the certificate must
match a whitelist.

The certificate security mode is not enabled by default, and instead
IP-based security mode is used. It resolves the same whitelist as
hostnames into IPs to make an IP whitelist of allowed clients.

StartUp was required to become a ServletContentListener so that it could
initialize Spring by the time the HostnameSecurity filter started.
StartUp works better as a listener anyway.

Code review: http://codereview.appspot.com/5607044/