Fix file system connector code site issue 32: Initial snapshot fails in Java 7.
http://code.google.com/p/google-enterprise-connector-file-system/issues/detail?id=32

The custom Comparator expressly failed on two equal arguments, which
does not meet the specification for a Comparator and fails with
TreeMap (and hence TreeSet) in Java 7, which compares the first
inserted key with itself.

http://bugs.sun.com/view_bug.do?bug_id=5045147

Duplicate snapshot numbers are not actually possible here, so I just
removed the check. I also removed an extraneous filter on the snapshot
directory (there were two), and changed a call to avoid autoboxing.

Code review: http://codereview.appspot.com/6174045/