Fix NullPointerExceptions thrown out of ChangeListener in ConnectorCoordinatorTest
Some of the tests in ConnectorCoordinatorTest would throw NPE out
of the ChangeListener.  The problem was almost always null configurations,
schedules, etc read out of the persistent store when adding/updating
a connector.  After examining many, many failures, they all seemed to
be in or after testThreadDeadlock().  Apparently one or more of the 
threads created by each call to checkThreadDeadlock() were not quite
finished when exiting that routine, and continued stomping on the 
PStore.  

Joining the two threads a the bottom of checkThreadDeadlock() seems
to have cleared up the problem.  Now, the other checks in that
method, Thread.isAlive(), would seem to indicate that the threads
should already be finished; but closer examination shows we break
out of the loop if one finishes before the other.

I do believe there is a problem with ChangeDetector.  In a true
HA environment, it is quite possible that the schedule could be
saved to the persistent store before the configuration, leading
to the same situation.  Pehaps it only happens if the configuration
is deleted asychronously, so the ChangeDetector sees a configuration,
but ChangeListener doesn't.  I believe the testThreadDeadlock()
test bashes on the persistent store much harder than any real-life
scenerio, and these tests exposed windows of inconsistency.


Change Log:
----------
M projects/connector-manager/source/javatests/com/google/enterprise/connector/instantiator/ConnectorCoordinatorTest.java
M projects/connector-manager/source/javatests/com/google/enterprise/connector/persist/MockPersistentStore.java

Code Review:
-----------
http://codereview.appspot.com/6158044