Fix issue 7928861: Avoid unordered snapshots in the snapshot files.
This fix is pedantically correct; the diffing package fundamentally
requires ordered docids. However, for certain scenarios, this change
will lead to additional documents being repeatedly treated as new
documents. In most of those scenarios, however, that behavior is
instead of the current behavior of permanently deleting documents that
actually exist. The problems occur when the docids from the repository
are sorted in a non-lexicographic order, since the diffing package
using string comparisons that are lexicographic.

To make that easier to fix in diffing connectors, the
DocumentSnapshotRepositoryMonitor now supports DocumentSnapshot classes
that implement Comparable, and uses compareTo on the snapshots
themselves instead of comparing the document IDs. Connectors can
implement a natural ordering without having to produce a string
document ID whose lexicographic ordering matching the desired natural
ordering. Connectors that do not implement Comparable on their
snapshots will continue to rely on the lexicographic ordering of the
document IDs.

For full correctness in the database connector, this depends on a fix
for b/7732455 (probably now implementing a Comparable DocumentSnapshot).

Other changes:

* Changed "file system" to "repository" in doc comments.
* Fixed @Override annotations, removing erroneous ones, adding missing
  ones, and uncommenting them all.
* Fixed an error message and related TODO comment typos.
* Removed unused imports in MockDocumentSnapshot, which had been
  otherwise modified in patch set 1.
* Reordered the traversalSchedule field below the statics.

Code review: https://codereview.appspot.com/7030044/