Fix issue 7040140 CM should provide Last-Modified date in getDocumentContent
This change provides a Last-Modified date in HttpResponse from the
GetDocumentContent servlet.  For historical reasons the servlet
did not provide the Last-Modified header because earlier versions
of the GSA ignored it.

This change implements getLastModfied() and lets the stock
HttpServlet code handle If-Modified-Since and Last-Modified support.
It removes our custom If-Modified-Since handling.

Code Review:
http://codereview.appspot.com/7129047