Fix issue 8237465: Connector instantiation fails with non-ASCII characters in the advanced properties XML.
The XML document was converted to bytes using the platform default
character encoding, but the XML declaration claimed it was UTF-8.

I converted ByteArrayResourceHack to a method for brevity and added
support for a character encoding, using UTF-8 for the XML instance and
PropertiesUtils.PROPERTIES_ENCODING (Latin-1) for properties (to be
pedantic; it's really ASCII by the time we get the string, see
java.util.Properties for details).

Code review: https://codereview.appspot.com/7382044/