Log exceptions thrown out of connector's checkpoint method.
For historical reasons, QueryTraverser.checkpointAndSave()
would swallow exceptions thrown out of the connector's
checkpoint() method.  This was to accomodate early 1.x
connectors after we changed the DocumentList semantics to
allow empty document lists and improved exception handling.

However, a customer is running into a problem where the
Sharepoint connector is stuck in a loop, feeding the same
documents repeatedly.  This is because the Sharepoint
connector's checkpoint() method is throwing a RuntimeException.
Troubleshooting the problem is made more difficult because
the RuntimeException is swallowed silently by the 
Connector Manager.

Code Review:
http://codereview.appspot.com/8839048