Fix issue 9415184: FS connector gets blocked on I/O in MimeTypeDetector.
MimeUtil2 is not thread-safe and leaks memory, so we synchronize access
to a single instance. We tried to avoid letting it do file I/O, but a
customer found the connector hanging inside the MagicMimeDetector and
blocking all other retriever threads waiting for the synchronization
lock.

To avoid the problem, this change uses two separate instances, one
using only the extension detector which we give the file name to, and
the other using the magic detector which we give only the byte[] to.
The extension detector does not try to read files, and the magic
detector has no file to read.

The unit test is a little fragile. It doesn't actually fail on the old
code or the new code, but it would fail on the new code if
MimeTypeDetector used the actual file content when it wasn't supposed
to. It fails on the old code with just the "/dev/null/" prefix removed.

Code review: https://codereview.appspot.com/10238044