Remove log-and-throw anti-pattern from EncryptedPropertyPlaceholderConfigurer.
There was an editing snafu in r2772 that duplicated a logging call
(the error is not present in the code review). This code was also
edited in r2705 to improve the message and log the failed property
name, and trivially in r2505. The log-and-throw was original in r606.

The log-and-throw just duplicates the log messages, and the typo means
the message is logged three times, once with a stack trace, but unless
the logging is at FINEST, the original exception is not logged at all.
Since we're going to log the stack trace when the exception is caught
anyway, I just pass along the original exception as the cause.

Other changes:

* Removed public access from loadProperties and storeProperties in
  PropertiesUtils. These methods are only used internally and by tests.

Code review: http://codereview.appspot.com/41570043/