Diffing "Too many recovery files." exception accumulates recovery files.
This change has CheckpointAndChangeQueue.resume() remove the recovery
file it created if an IOException was thrown writing it or removing
excess recovery state.  This does not attempt to fix whatever is 
causing recovery state to be corrupt, but it does try to avoid 
filling the disk with recovery files.

Code Review: http://codereview.appspot.com/103200044